chore(): Unity updates for the close button with delay updates (#530)
* chore(): Unity updates for the close button with delay updates

* chore(): Unity Android public API cleanup

* Update saunity/src/main/java/tv/superawesome/plugins/publisher/unity/SAUnityVideoAd.java


* Update saunity/src/main/java/tv/superawesome/plugins/publisher/unity/SAUnityInterstitialAd.java


* Update saunity/src/main/java/tv/superawesome/plugins/publisher/unity/SAUnityVideoAd.java


* chore(): Unity Android close button with custom delay updates

* chore(): Unity close button updates

* chore(): spacing fix

* chore(): unit test fixes

* chore(): fixing double multiplication of the close button delay

* Update saunity/src/main/java/tv/superawesome/plugins/publisher/unity/SAUnityVideoAd.java


* Update saunity/src/main/java/tv/superawesome/plugins/publisher/unity/SAUnityInterstitialAd.java


* Update saunity-common/src/main/java/tv/superawesome/plugins/publisher/unity/SAUnityVideoAd.java


* Update saunity-common/src/main/java/tv/superawesome/plugins/publisher/unity/SAUnityInterstitialAd.java


---------

+++ superawesome-base/src/main/java/tv/superawesome/sdk/publisher/managed/ManagedAdConfig.kt
@@ -39,7 +39,7 @@ internal class ManagedAdConfig : Parcelable {
         shouldShowCloseWarning = input.readByte().toInt() != 0
         isBackButtonEnabled = input.readByte().toInt() != 0
         autoCloseAtEnd = input.readByte().toInt() != 0
-        closeButtonState = CloseButtonState.fromInt(input.readInt(), input.readLong())
+        closeButtonState = CloseButtonState.fromInt(input.readInt(), input.readDouble())
         environment = SAConfiguration.fromOrdinal(input.readInt())
     }
 
@@ -52,7 +52,7 @@ internal class ManagedAdConfig : Parcelable {
         parcel.writeByte((if (isBackButtonEnabled) 1 else 0).toByte())
         parcel.writeByte((if (autoCloseAtEnd) 1 else 0).toByte())
         parcel.writeInt(closeButtonState.value)
-        parcel.writeLong(closeButtonState.time)
+        parcel.writeDouble(closeButtonState.time)
         parcel.writeInt(environment.ordinal)
     }
